refactor: move statistics into a separate component
